feat: v0.13.0 - false positive reduction, fix guidance, TS6, vitest 4
False positive reduction:
- Known example keys excluded (AWS AKIAIOSFODNN7EXAMPLE, etc.)
- Placeholder indicators detected (example, placeholder, your_, fake_, etc.)

Scan fix guidance:
- Each finding includes actionable fix text with rotation URLs

Refactor:
- Split commands/secrets.ts (409) into secrets.ts (234) + env-run.ts (177)

Dependencies:
- TypeScript 6.0.2, vitest 4.1.2, 0 npm audit vulnerabilities
- tsconfig: module/moduleResolution set to Node16

804 tests pass.
