+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "secretless-ai",
-  "version": "0.12.7",
+  "version": "0.13.0",
   "description": "One command to keep secrets out of AI. Works with Claude Code, Cursor, Copilot, Windsurf, and any AI coding tool.",
   "license": "Apache-2.0",
   "type": "commonjs",
@@ -45,7 +45,7 @@
   },
   "devDependencies": {
     "@types/node": "^25.5.0",
-    "typescript": "^5.9.3",
-    "vitest": "^2.1.9"
+    "typescript": "^6.0.2",
+    "vitest": "^4.1.2"
   }
 }
