+++ src/cli.ts
@@ -9,7 +9,8 @@ import * as path from 'path';
 import { VERSION } from './commands/utils';
 import { runInit, runScan, runStatus, runVerify, runDoctor } from './commands/core';
 import { runClean, runWatch, runScanHistory, runCleanHistory } from './commands/transcript';
-import { runSecret, runRun, runEnv, runImport, runSetupCommand } from './commands/secrets';
+import { runSecret } from './commands/secrets';
+import { runRun, runEnv, runImport, runSetupCommand } from './commands/env-run';
 import { runRules } from './commands/rules';
 import { runHook, runScanStaged } from './commands/git';
 import { runProtectMcp, runMcpStatus, runMcpUnprotect } from './commands/mcp';
